ParkingTicket.h

+++ ParkingTicket.h
@@ -7,14 +7,14 @@ class ParkingTicket {
 private:
     int ticketNumber;
     bool paid;
-    int amount;
+    double amount;
 public:
     int duration;
 public:
     ParkingTicket(int number);
     int getTicketNumber() const;
     void calPrice(int duration) ; 
-    int getPrice();
+    double getPrice();
     bool isPaid() const;
     void markAsPaid();
 };
